It has become evident that different target packages are needed:
This is because in order for SSR (with React) to work properly in the type system we need DOM definitions -- something I'm not willing to provide for **every** js target.

Here's an overview of targets:

- js-none: projects that don't use any browser or node APIs
- js-web: projects that use browser APIs
- js-node: projects that use node APIs
- js-hybrid: projects that both use node and browser APIs

If JSX is used, the jsx-* variant should be used instead. (JSX is handled separately because it requires react)