Remove jsx-none and jsx-node.
JSX implies React and React is inherently tied to the browser DOM:

It is not possible to use react-dom in a node environment without also having to install DOM type definitions.

jsx-none might be doable but there isn't really a compelling reason to use it over jsx-web.

Therefore I'm removing the two to make things a little simpler.